+++ src/main/java/se/bjurr/prnfs/settings/PrnfsSettings.java
@@ -8,10 +8,15 @@ import java.util.List;
 public class PrnfsSettings {
  private List<PrnfsNotification> notifications = newArrayList();
  private final List<PrnfsButton> buttons;
+ private final boolean usersAllowed;
+ private final boolean adminsAllowed;
 
- public PrnfsSettings(List<PrnfsNotification> notifications, List<PrnfsButton> buttons) {
+ public PrnfsSettings(List<PrnfsNotification> notifications, List<PrnfsButton> buttons, boolean usersAllowed,
+   boolean adminsAllowed) {
   this.notifications = checkNotNull(notifications);
   this.buttons = checkNotNull(buttons);
+  this.usersAllowed = usersAllowed;
+  this.adminsAllowed = adminsAllowed;
  }
 
  public List<PrnfsNotification> getNotifications() {
@@ -20,5 +25,13 @@ public class PrnfsSettings {
 
  public List<PrnfsButton> getButtons() {
   return buttons;
+ }
+
+ public boolean isUsersAllowed() {
+  return usersAllowed;
+ }
+
+ public boolean isAdminsAllowed() {
+  return adminsAllowed;
  }
 }